+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK). Also see the [releases](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK/releases) on Github.
 ## Unreleased
+- [CM-842] Added Support for s3 service for upload & download
 - [CM-825] Fixed SPM integration issue
 ## [0.1.2] 2022-01-20
 - Fixed the Stroyboard linking issue